fix: use latest config in traceSendCommand
